Records list now accepts any iterable, not only an array
Previously there were several calls to `input.map` which work for arrays but not for generators or most other iterable objects. The code has been changed to use a `for(x of y)` loop instead of `map` functions. (As a side-effect, the new code is also more performant and memory-efficient!)